Moved dictionary to overlay. Now barely 740 bytes left there, but more than 3K on normal memory.
Also added SndStop to sound.s and tweaked so music is stopped correctly. Also added jsr SndStop before restarting to see if I can get the music be correctly played after restarting.

If music gets much higher than which currently is, be careful to remove dict again.

Found a bug, where the value of shields gets <0 or something like that and is not decremented afterwards..