First tweaks after the BETA1 version
Changed the way rear view is patched, so no it is in obj3d.s before calling GLOBROT. Is much more efficient.

Also corrected an error with this, as only Z component was inverted, and not also X component. This also meant changes into stars.s which also had some errors in rear view...

Tweaked the rotation of the initial screen ships, and added the hermit as type of ship (but not yet created in universe.s)