Added the outtro desk picture to the data and modified the outro display code to show it with the nice border/framing the main game is using.
Moved the arrow display code in another module to allow the outro to not display the navigation arrow block.

Added a define to skip the game entirelly and directly jump to the outro sequence.

The whole thing is super messy and will require some cleaning at some point, but at least it makes the development possible :D