Moved the mapping tables from game_data.c to game_text.s to simplify the code and avoid having to "extern" everything.
Fixed a problem with the chemistry recipies appearing at the wrong location if the chemistry book is read out of the library.

Moved the vocabular enum to the game_enums.h so they can be used from assembler (as well as C).

Simplified the way the container type is checked by using the property that they are the first items in the list, so no need to hardcode a switch case on specific values.