Did a large refactoring of the locations: - Removed the "description" field from the location structure - Added a SET_DESCRIPTION script command to provide the description. A benefit of that, is that now all the data for a location is grouped in a single location, with all the texts, images, logic, etc... instead of being spread over multiple locations in the code.
This looks kind of more messy, but that makes it much easier compare the translations options of the various languages, and makes it easier to dynamically change the description of a room depending of changing events.

A drawback is that now the "debugger" can't print the location description anymore.

Also fixed a problem with the InputMethod and the handling of the SHIFT key.