Cleaned up the asset build process by wrapping the calls to PictConv into a single "Convert" call, which required a few names changes so source files and target files actually matched.
Also modified a couple .s files which had to be manually copied into the source code by now having the converted file directly included, which means the PNG file can be edited and the change will be reflected.

And finally actually gave a name to each of the location files instead of the number from 1 to 40 that I inherited from the BASIC version of the game.

Also fixed an issue in the build script which would silently fail if a source file was missing: Now the output is printed in red when that happens.