Added the action to break the window with the cue.
Added the action to combine the window with the rope.

Added the action for the girl to climb down the rope.

Added the girl sitting on the small wall outside waiting for us and saying "thanks" and that she will follow us.

Removed a unused section with the location which was not in use anymore.