Hacked horribly the keyboard mapping into the game... should have decoupled the display of the layout from the actual code, but I just need for it to work at the moment, ideally I should move the keyboard code to high memory since it is used by all the modules anyway.
The result is that now whatever layout is selected in the splash (or saved) will be applied in both the game and the outro (when entering the high score).

Note: I forgot to handle the M character, but not sure where to map it, probably where ;: is located on the Oric keyboard?