Added support for defining the maximum size of the text that can be edited in the input system.
Also changed the default so the text is input as lower case and shift shows it as upper case (and fixed the logic in the token matcher so the casing is ignored).

Also change the location of the input field in the game, and the "What do you want to do now" message only appears once at the start of the game.