Added the possibility to check for which type a container an item is placed in, then modified the script to code to ensure that we can't create a bomb in a plastic bag or cardboard box.
Also fixed the set item location command to make sure it drops the contained item if a non empty container is dropped (or thrown).

Modified the gun powder so it can be transported in any adequate (for powder) container.

Made sure that using the wrong container shows a message telling the user they should use something more adapted.